[otp, rom, sw] Align with unified bypass and FIPS threshold registeres
This commit re-aligns the ROM and software implementations with the
latest implementation of the entropy_src RTL which has unified
threshold registers for FIPS / CC compliant and boot-time / bypass mode.

Because the hardware now only contains a single configuration at any
point in time:
1. ROM/software needs to select the proper configuration from OTP.
2. OTP needs to maintain separate configuration digest values for each
   mode.

In addition, this commit also introduces a separate alert threshold
configuration for both modes in OTP. Because the modes behave quite
differently, it may be desirable to configure different alert
thresholds.

+++ sw/device/lib/dif/dif_entropy_src.h
@@ -224,18 +224,16 @@ typedef struct dif_entropy_src_health_test_config {
    */
   dif_entropy_src_test_t test_type;
   /**
-   * The high threshold for the health test (contains both FIPS and bypass
-   * thresholds).
+   * The high threshold for the health test.
    */
-  uint32_t high_threshold;
+  uint16_t high_threshold;
   /**
-   * The low threshold for the health test (contains both FIPS and bypass
-   * thresholds).
+   * The low threshold for the health test.
    *
    * If the corresponding health test has no low threshold, set to 0, otherwise
    * `dif_entropy_src_health_test_configure()` will return `kDifBadArg`.
    */
-  uint32_t low_threshold;
+  uint16_t low_threshold;
 } dif_entropy_src_health_test_config_t;
 
 /**
